+++ ciprs_reader/parser/lines.py
@@ -105,36 +105,6 @@ class OffenseDispositionMethod(Parser):
         # report[self.state["section"]][-1]["Disposition Method"] = matches["value"]
 
 
-class DefendantParser(Parser):
-    """Only enabled when in Defendant section."""
-
-    def is_enabled(self):
-        return self.state.section == Section.DEFENDANT
-
-
-class DefendantRace(DefendantParser):
-
-    pattern = r"\s*Race: \s*(?P<value>\w+)"
-    section = ("Defendant", "Race")
-
-
-class DefendantSex(DefendantParser):
-
-    pattern = r"\s*Sex: \s*(?P<value>\w+)"
-    section = ("Defendant", "Sex")
-
-    def clean(self, matches):
-        """Parse and convert defendent sex to M or F"""
-        sex = matches["value"].lower()
-        if sex == "female":
-            matches["value"] = "F"
-        elif sex == "male":
-            matches["value"] = "M"
-        else:
-            matches["value"] = ""
-        return matches
-
-
 class DefendentDOB(Parser):
 
     pattern = r"\s*Date of Birth/Estimated Age:[\sa-zA-Z]*(?P<value>[\d/]+)[ ]{2,}"
