+++ ciprs_reader/parser/section/defendant.py
@@ -1,0 +1,32 @@
+from ciprs_reader.const import Section
+from ciprs_reader.parser.base import Parser
+
+
+class DefendantParser(Parser):
+    """Only enabled when in Defendant section."""
+
+    def is_enabled(self):
+        return self.state.section == Section.DEFENDANT
+
+
+class DefendantRace(DefendantParser):
+
+    pattern = r"\s*Race: \s*(?P<value>\w+)"
+    section = ("Defendant", "Race")
+
+
+class DefendantSex(DefendantParser):
+
+    pattern = r"\s*Sex: \s*(?P<value>\w+)"
+    section = ("Defendant", "Sex")
+
+    def clean(self, matches):
+        """Parse and convert defendent sex to M or F"""
+        sex = matches["value"].lower()
+        if sex == "female":
+            matches["value"] = "F"
+        elif sex == "male":
+            matches["value"] = "M"
+        else:
+            matches["value"] = ""
+        return matches
